Centralized Java version parsing
