+++ sandbox/build.gradle
@@ -9,6 +9,7 @@ dependencies {
     api project(":utils")
     api project(":shadowapi")
     api project(":utils:reflector")
+    api project(":utils:jvm")
     compileOnly "com.google.auto.service:auto-service-annotations:1.0-rc6"
     api "javax.annotation:javax.annotation-api:1.3.2"
     api "javax.inject:javax.inject:1"
