+++ settings.gradle
@@ -5,6 +5,7 @@ include ":sandbox"
 include ":junit"
 include ":utils"
 include ":utils:reflector"
+include ":utils:jvm"
 include ":pluginapi"
 include ":plugins:accessibility-deprecated"
 include ":plugins:maven-dependency-resolver"
