+++ utils/reflector/src/main/java/org/robolectric/util/reflector/UnsafeAccess.java
@@ -6,6 +6,8 @@ import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.security.ProtectionDomain;
+
+import org.robolectric.util.JavaVersion;
 import sun.misc.Unsafe;
 
 /** Access to sun.misc.Unsafe and the various scary things within. */
@@ -13,7 +15,7 @@ import sun.misc.Unsafe;
 public class UnsafeAccess {
 
   private static final Danger DANGER =
-      getJavaVersion() < 11
+      JavaVersion.SYSTEM < 11
           ? new DangerPre11()
           : new Danger11Plus();
 
@@ -99,27 +101,5 @@ public class UnsafeAccess {
         throw new AssertionError(e);
       }
     }
-  }
-
-  /**
-   * Returns the Java version as an int value.
-   *
-   * @return the Java version as an int value (8, 9, etc.)
-   */
-  private static int getJavaVersion() {
-    String version = System.getProperty("java.version");
-    assert version != null;
-    if (version.startsWith("1.")) {
-      version = version.substring(2);
-    }
-    // Allow these formats:
-    // 1.8.0_72-ea
-    // 9-ea
-    // 9
-    // 9.0.1
-    int dotPos = version.indexOf('.');
-    int dashPos = version.indexOf('-');
-    return Integer.parseInt(
-        version.substring(0, dotPos > -1 ? dotPos : dashPos > -1 ? dashPos : 1));
   }
 }
