+++ utils/src/main/java/org/robolectric/util/Util.java
@@ -17,28 +17,6 @@ import java.util.List;
  */
 public class Util {
 
-  /**
-   * Returns the Java version as an int value.
-   *
-   * @return the Java version as an int value (8, 9, etc.)
-   */
-  public static int getJavaVersion() {
-    String version = System.getProperty("java.version");
-    assert version != null;
-    if (version.startsWith("1.")) {
-      version = version.substring(2);
-    }
-    // Allow these formats:
-    // 1.8.0_72-ea
-    // 9-ea
-    // 9
-    // 9.0.1
-    int dotPos = version.indexOf('.');
-    int dashPos = version.indexOf('-');
-    return Integer.parseInt(
-        version.substring(0, dotPos > -1 ? dotPos : dashPos > -1 ? dashPos : version.length()));
-  }
-
   public static void copy(InputStream in, OutputStream out) throws IOException {
     byte[] buffer = new byte[8196];
     int len;
